+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Bugfixes:
 * Fix flatpak package count (#441)
 * Don't print white color blocks with `--pipe` (#450)
 * Fix iTerm being detected as iTermServer-* sometimes
+* Fix sound device volume being incorrectly detected as muted sometimes (Sound)
 
 Logo:
 * Update Windows 11 ASCII logo to look more visually consistent (#445)